+++ libs/gesha-core/src/conversions/v3_0/to_openapi/mod.rs
@@ -1,4 +1,9 @@
 mod to_components_object;
 mod to_document;
 mod to_paths_object;
+
 mod to_request_body;
+use to_request_body::to_request_body_pair;
+
+mod to_schema_case;
+use to_schema_case::to_schema_pair;
