+++ libs/openapi-types/src/v3_0/media_type.rs
@@ -1,3 +1,5 @@
+use crate::v3_0::SchemaCase;
+
 /// > The key is a media type or media type range and the value describes it.
 #[derive(Clone, Debug, Hash, Eq, PartialEq)]
 pub struct MediaTypeKey(String);
@@ -8,5 +10,8 @@ impl MediaTypeKey {
     }
 }
 
+/// rf. https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#media-type-object
 #[derive(Clone, Debug)]
-pub struct MediaTypeObject {}
+pub struct MediaTypeObject {
+    pub schema: Option<SchemaCase>,
+}
